+++ src/Components/Search/Previews/index.tsx
@@ -1,5 +1,5 @@
 import React, { SFC } from "react"
-import { ArtistSearchPreviewQueryRenderer as ArtistSearchPreview } from "./Grids/Artist"
+import { ArtistSearchPreviewQueryRenderer as ArtistSearchPreview } from "./Grids/ArtistSearch"
 import { MerchandisableArtworksPreviewQueryRenderer as MerchandisableArtworksPreview } from "./Grids/MerchandisableArtworks"
 
 export interface SearchPreviewProps {
